chore: cleaned up some unsigned int-related overflow issue in vsetMakers chore: changed Node, Neighbor and Element accessors in Element, Face, and InterFace removing const-ness from objects returned via pointers because neither of the classes owns them chore: cleaned up various leftover occurrences of size_t that were causing warnings

+++ source/includes/model/CSMP_global_enumerations.h
@@ -92,7 +92,7 @@ enum PLACEMENT : std::int8_t { UNDEFINED, // default
 enum CELL_SHAPE : std::int8_t { LINE=1, SURFACE=2, VOLUME=3, HYPER_DIMENSIONAL=4 };
 
 /// Side of lower-dimensional face or interface between two higher-dimensional elements ( INSIDE or OUTSIDE ) and a potential lower-dimensional parent element ( MIDDLE )
-enum INTERFACE_SIDE : std::int8_t
+enum INTERFACE_SIDE : std::int32_t
 {
     INSIDE  = -1,
     OUTSIDE =  1,
